Move utils/config.js to config/index.js
Since the config/ directory has been created for sequelize, we might as well use
the directory for the actual configs instead of just placing them in an "utils"
directory.
