--- config/config.json
@@ -1,30 +1,0 @@
-{
-  "development": {
-    "username": "postgres",
-    "database": "dev_postgres",
-    "port": "5432",
-    "host": "db",
-    "dialect": "postgres",
-    "operatorsAliases": false,
-    "pool": {
-      "max": 5,
-      "min": 0,
-      "acquire": 10000,
-      "idle": 300000000
-    }
-  },
-  "production": {
-    "username": "postgres",
-    "database": "postgres",
-    "port": "5432",
-    "host": "db",
-    "dialect": "postgres",
-    "operatorsAliases": false,
-    "pool": {
-      "max": 5,
-      "min": 0,
-      "acquire": 10000,
-      "idle": 300000000
-    }
-  }
-}
